chore: move code under pkg
